pyproject.toml
[project]
name = "tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "polars>=1.17.1",
    "pyarrow>=18.1.0",
    "pyyaml>=6.0.2",
    "ruff>=0.8.3",
    "scipy>=1.14.1",
    "streamlit>=1.41.1",
    "uvicorn>=0.32.1",
    "yfinance>=0.2.50",
]
